remove readPreferences file, move to advanced-tab
